Added encoding operations
